--- Phase04/SimpleCalculator-master/SimpleCalculator.Business/Enums/OperatorEnum.cs
@@ -1,10 +1,0 @@
-﻿namespace SimpleCalculator.Business.Enums
-{
-    public enum OperatorEnum
-    {
-        sum,
-        sub,
-        multiply,
-        division
-    }
-}
