--- Phase04/SimpleCalculator-master/SimpleCalculator.Business/OperatorBusiness/Operators/MultiplyOperator.cs
@@ -1,12 +1,0 @@
-﻿using SimpleCalculator.Business.Abstraction;
-
-namespace SimpleCalculator.Business.OperatorBusiness.Operators
-{
-    internal class MultiplyOperator : IOperator
-    {
-        public int Calculate(int first, int second)
-        {
-            return first * second;
-        }
-    }
-}
