--- Phase04/SimpleCalculator-master/SimpleCalculator/Program.cs
@@ -1,5 +1,0 @@
-﻿using SimpleCalculator.Business;
-using SimpleCalculator.ConsoleApp;
-
-new UiManager(new Calculator())
-    .StartUI();
