+++ Phase04/SimpleCalculator.Business/Enums/OperatorEnum.cs
@@ -1,0 +1,10 @@
+﻿namespace Phase04.SimpleCalculator.Business.Enums
+{
+    public enum OperatorEnum
+    {
+        Sum,
+        Sub,
+        Multiply,
+        Division
+    }
+}
