+++ Phase04/SimpleCalculator/Program.cs
@@ -1,0 +1,5 @@
+﻿using Phase04.SimpleCalculator.Business;
+using Phase04.SimpleCalculator;
+
+new UiManager(new Calculator())
+    .StartUI();
